fixed volume max value being out of bounds but FOR REAL THIS TIME

+++ src/tetris/settings/GameSettings.java
@@ -5,7 +5,7 @@ import tetris.music.MusicPlayer;
 
 public class GameSettings {
     public int music; // 0 for no music, 1 for NightSnow, 2 for Vremya, 3 for VirtualLight
-    public int volume; // self-explanatory
+    public double volume; // self-explanatory
 
     public int renderFPS; // the number of frames per second to render at
